Support overriding options when calling magit-imerge-finish
When finishing an incremental merge, the initial values for the branch
and goal options can be overridden, but magit-imerge-finish doesn't
support this because it's a sequence action and doesn't have access to
the options from the main imerge popup.  Add a sequence action for a
second imerge popup that stores the arguments for later use by
magit-imerge-finish.