status: Don't highlight equivalent options as overriding
We highlight any option set through magit-imerge-set-finish-arguments
as an overriding one.  While an option most likely is overriding the
current value because the user is bothering to set it (the popup
options aren't populated with their current value), we shouldn't
highlight an option that isn't actually changing.

This will become more important later because we will populate the
popup (or actually transient, at that point) with the current values,
even after a merge is resumed from its state blob.