Convert popups to transients
First do the straightforward conversion.  The next commits will move
the magit-imerge-finish-options transient into the main magit-imerge
transient.  The reason that the secondary popup existed was to allow
changing the options of the current merge while the main popup was in
its sequencer state.  Transients do not have this limitation.